+++ tests/fixtures/explicit_background_td_db_fixture.py
@@ -4,7 +4,7 @@ import bw2data as bd
 import numpy as np
 import pytest
 from bw2data.tests import bw2test
-from bw_temporalis import TemporalDistribution
+from bw_timex import TemporalDistribution
 
 
 @pytest.fixture