+++ tests/test_temporal_evolution.py
@@ -80,7 +80,7 @@ class TestMutualExclusivity:
 import bw2data as bd
 import numpy as np
 from bw2data.tests import bw2test
-from bw_temporalis import TemporalDistribution
+from bw_timex import TemporalDistribution
 
 from bw_timex import TimexLCA